Role permissions1

+++ app/Http/Controllers/Dashboard/CategoryController.php
@@ -9,6 +9,8 @@ use App\Models\Category;
 use App\Traits\FileMethods;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Gate;
+
 
 
 
@@ -24,6 +26,9 @@ class CategoryController extends Controller
     }
     public function index(Request $request)
     {
+        if (!Gate::allows('categories.view')){
+            abort(403, 'Unauthorized');
+        } 
 
         // SELECT a.*, b.name as parent_name FROM categories as a
         // LEFT JOIN cateories as b ON b.id = a.parent_id
@@ -46,6 +51,7 @@ class CategoryController extends Controller
 
     public function create()
     {
+        Gate::authorize('categories.create');
         $category = new Category();
         $parents = $this->categoryRepository->getAllCategories();
         return view("dashboard.categories.create", compact("parents", "category"));
@@ -53,6 +59,8 @@ class CategoryController extends Controller
 
     public function store(CategoryRequest $request)
     {
+        // Gate::authorize('categories.create');
+
         $data = array_merge(
             $request->except('image'),
             ['slug' => Str::slug($request->name)]
@@ -67,11 +75,14 @@ class CategoryController extends Controller
 
     public function show(Category $category)
     {
+        Gate::authorize('categories.view');
+
        return view("dashboard.categories.show", compact("category"));
     }
 
     public function edit($CategoryId)
     {
+        Gate::authorize('categories.edit');
         $category = $this->categoryRepository->getCategoryById($CategoryId);
         $parents = Category::where('id', '!=', $CategoryId)
             ->where(function ($query) use ($CategoryId) {
@@ -84,6 +95,7 @@ class CategoryController extends Controller
 
     public function update(CategoryRequest $request, $CategoryId)
     {
+        // Gate::authorize('categories.edit');
         $data = array_merge($request->except('image', '_token', '_method', 'id'), ['slug' => Str::slug($request->name)]);
         $category = $this->categoryRepository->getCategoryById($CategoryId);
         if ($request->hasFile('image')) {
@@ -100,6 +112,8 @@ class CategoryController extends Controller
 
     public function destroy($CategoryId)
     {
+        Gate::authorize('categories.delete');
+
         // $imagePath = Category::findOrFail($CategoryId)->image;
         $this->categoryRepository->deleteCategory($CategoryId);
         // $this->deleteFile($imagePath);
